creates addPair.js
This module stores a function that takes on three variables: the class of the elements to assign a click handler (`listeningEl`), and the variant ids for the left and right (leftVariant, rightVariant).

On click uses a fetch call to wait for both variants to be loaded into the cart before moving to the cart page.